# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Uttam Kirana - Grocery Shop Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface SendOtpBody {
  mobile: string;
}

export interface OtpResponse {
  message: string;
  /** Simulated OTP (returned in dev mode) */
  otp?: string;
}

export interface VerifyOtpBody {
  mobile: string;
  otp: string;
  name?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  admin: "admin",
  delivery: "delivery",
} as const;

export interface User {
  id: number;
  mobile: string;
  name?: string;
  role: UserRole;
  address?: string | null;
  city?: string | null;
  pincode?: string | null;
  walletBalance: number;
  rewardPoints: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateAddressBody {
  address: string;
  city?: string;
  pincode?: string;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  icon?: string;
  color?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon?: string | null;
  color?: string | null;
}

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  mrp?: number | null;
  unit: string;
  stock: number;
  categoryId: number;
  categoryName?: string | null;
  imageUrl?: string | null;
  isActive: boolean;
  isFeatured: boolean;
  discount?: number | null;
  createdAt: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
}

export interface CreateProductBody {
  name: string;
  description?: string;
  price: number;
  mrp?: number;
  unit: string;
  stock: number;
  categoryId: number;
  imageUrl?: string;
  isFeatured?: boolean;
  discount?: number;
}

export interface UpdateProductBody {
  name?: string;
  description?: string;
  price?: number;
  mrp?: number;
  unit?: string;
  stock?: number;
  categoryId?: number;
  imageUrl?: string;
  isActive?: boolean;
  isFeatured?: boolean;
  discount?: number;
}

export interface UpdateStockBody {
  stock: number;
}

export interface CartItem {
  productId: number;
  productName: string;
  price: number;
  imageUrl?: string | null;
  unit: string;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  itemCount: number;
  subtotal: number;
  deliveryCharge: number;
  total: number;
}

export interface CartItemBody {
  productId: number;
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  price: number;
  quantity: number;
  subtotal: number;
  unit: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  placed: "placed",
  confirmed: "confirmed",
  packed: "packed",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cod: "cod",
  online: "online",
} as const;

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
} as const;

export interface Order {
  id: number;
  userId: number;
  userName?: string | null;
  userMobile?: string | null;
  deliveryBoyId?: number | null;
  deliveryBoyName?: string | null;
  items: OrderItem[];
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  paymentStatus: OrderPaymentStatus;
  subtotal: number;
  deliveryCharge: number;
  total: number;
  deliveryAddress: string;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  cod: "cod",
  online: "online",
} as const;

export interface CreateOrderBody {
  paymentMethod: CreateOrderBodyPaymentMethod;
  deliveryAddress: string;
  notes?: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  placed: "placed",
  confirmed: "confirmed",
  packed: "packed",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface AssignDeliveryBody {
  deliveryBoyId: number;
}

export interface AnalyticsSummary {
  totalOrders: number;
  totalRevenue: number;
  todayOrders: number;
  todayRevenue: number;
  totalCustomers: number;
  pendingOrders: number;
  outForDeliveryOrders: number;
  averageOrderValue: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalSold: number;
  revenue: number;
  imageUrl?: string | null;
}

export interface OrderStatusCount {
  status: string;
  count: number;
}

export type GetProductsParams = {
  search?: string;
  category?: string;
  inStock?: boolean;
  minPrice?: number;
  maxPrice?: number;
  limit?: number;
  offset?: number;
};

export type GetOrdersParams = {
  status?: string;
  limit?: number;
  offset?: number;
};

export type GetTopProductsParams = {
  limit?: number;
};