# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Uttam Kirana - Grocery Shop Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login/register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/me/address:
    put:
      operationId: updateAddress
      tags: [auth]
      summary: Update delivery address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAddressBody"
      responses:
        "200":
          description: Address updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Categories
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{categoryId}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  # Products
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear the cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemBody"
      responses:
        "200":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # Orders
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    SendOtpBody:
      type: object
      properties:
        mobile:
          type: string
      required: [mobile]
    OtpResponse:
      type: object
      properties:
        message:
          type: string
        otp:
          type: string
          description: Simulated OTP (returned in dev mode)
      required: [message]
    VerifyOtpBody:
      type: object
      properties:
        mobile:
          type: string
        otp:
          type: string
        name:
          type: string
      required: [mobile, otp]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    UpdateAddressBody:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
        pincode:
          type: string
      required: [address]
    User:
      type: object
      properties:
        id:
          type: integer
        mobile:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [customer, admin, delivery]
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        pincode:
          type: string
          nullable: true
        walletBalance:
          type: number
        rewardPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, mobile, role, walletBalance, rewardPoints, createdAt]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        color:
          type: string
      required: [name, slug]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
      required: [id, name, slug]
...
    CartItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        unit:
          type: string
        quantity:
          type: integer
        subtotal:
          type: number
      required: [productId, productName, price, unit, quantity, subtotal]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        itemCount:
          type: integer
        subtotal:
          type: number
        deliveryCharge:
          type: number
        total:
          type: number
      required: [items, itemCount, subtotal, deliveryCharge, total]
    CartItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
...